requirements/local.txt
# Add test and quality tools over here
-r base.txt

coverage==4.3.3
django-coverage-plugin==1.4

django-silk==0.7.2